requirements.txt
Flask==2.2.2
requests==2.28.1
python-dotenv==1.0.0  # Optional, if you're using .env for API keys